+++ templates/base_kit_app/lib/di/modules/di_module.dart
@@ -2,6 +2,7 @@ import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -10,9 +11,19 @@ import 'package:shared_preferences/shared_preferences.dart';
 @immutable
 abstract class DiModule {
   @preResolve
-  Future<SharedPreferences> get prefs => SharedPreferences.getInstance();
+  Future<SharedPreferences> sharedPreferences() async {
+    final getIt = GetIt.I;
+    if (getIt.isRegistered<SharedPreferences>()) {
+      return getIt<SharedPreferences>();
+    }
+    return SharedPreferences.getInstance();
+  }
 
   FlutterSecureStorage get secureStorage {
+    final getIt = GetIt.I;
+    if (getIt.isRegistered<FlutterSecureStorage>()) {
+      return getIt<FlutterSecureStorage>();
+    }
     return const FlutterSecureStorage(
       aOptions: AndroidOptions(encryptedSharedPreferences: true),
       iOptions: IOSOptions(),
@@ -20,7 +31,19 @@ abstract class DiModule {
   }
 
   @preResolve
-  Future<PackageInfo> get packageInfo => PackageInfo.fromPlatform();
+  Future<PackageInfo> packageInfo() async {
+    final getIt = GetIt.I;
+    if (getIt.isRegistered<PackageInfo>()) {
+      return getIt<PackageInfo>();
+    }
+    return PackageInfo.fromPlatform();
+  }
 
-  DeviceInfoPlugin get deviceInfo => DeviceInfoPlugin();
+  DeviceInfoPlugin deviceInfo() {
+    final getIt = GetIt.I;
+    if (getIt.isRegistered<DeviceInfoPlugin>()) {
+      return getIt<DeviceInfoPlugin>();
+    }
+    return DeviceInfoPlugin();
+  }
 }